backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pymongo==4.6.1
firebase-admin==6.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
dnspython==2.4.2

# AI Services
vosk==0.3.45
soundfile==0.12.1
gTTS==2.5.0
mediapipe==0.10.9
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
